move make controls to its own file
